wechat-bot/requirements.txt
flask>=3.0.0
requests>=2.31.0
discord.py>=2.3.0
